# Conflicts:
#	api-doc/nise-3-organization-api-service.yaml
#	app/Http/Controllers/OrganizationController.php
#	app/Models/BaseModel.php
#	app/Services/OrganizationService.php
#	database/factories/OrganizationFactory.php
#	database/factories/OrganizationUnitFactory.php
#	database/migrations/2021_05_11_100231_create_organizations_table.php
#	routes/web.php

+++ app/Http/Controllers/HumanResourceController.php
@@ -6,6 +6,7 @@ use App\Models\HumanResource;
 use App\Services\HumanResourceService;
 use Carbon\Carbon;
 use Exception;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Response;
@@ -37,14 +38,18 @@ class HumanResourceController extends Controller
      * @param Request $request
      * @return Exception|JsonResponse|Throwable
      * @throws ValidationException
+     * @throws AuthorizationException
+     * @throws Throwable
      */
     public function getList(Request $request): JsonResponse
     {
+        $this->authorize('viewAny', HumanResource::class);
+
         $filter = $this->humanResourceService->filterValidator($request)->validate();
         try {
             $response = $this->humanResourceService->getHumanResourceList($filter, $this->startTime);
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
 
         return Response::json($response);
@@ -54,13 +59,20 @@ class HumanResourceController extends Controller
      * Display the specified resource.
      * @param int $id
      * @return Exception|JsonResponse|Throwable
+     * @throws AuthorizationException
+     * @throws Throwable
      */
     public function read(int $id): JsonResponse
     {
         try {
             $response = $this->humanResourceService->getOneHumanResource($id, $this->startTime);
+            if (!$response) {
+                abort(ResponseAlias::HTTP_NOT_FOUND);
+            }
+            $this->authorize('view', $response['data']);
+
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response);
 
@@ -71,9 +83,12 @@ class HumanResourceController extends Controller
      * @param Request $request
      * @return Exception|JsonResponse|Throwable
      * @throws ValidationException
+     * @throws AuthorizationException|Throwable
      */
     public function store(Request $request): JsonResponse
     {
+        $this->authorize('create', HumanResource::class);
+
         $validatedData = $this->humanResourceService->validator($request)->validate();
         try {
             $data = $this->humanResourceService->store($validatedData);
@@ -87,7 +102,7 @@ class HumanResourceController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
 
         return Response::json($response, ResponseAlias::HTTP_CREATED);
@@ -99,10 +114,12 @@ class HumanResourceController extends Controller
      * @param int $id
      * @return Exception|JsonResponse|Throwable
      * @throws ValidationException
+     * @throws AuthorizationException|Throwable
      */
     public function update(Request $request, int $id): JsonResponse
     {
         $humanResource = HumanResource::findOrFail($id);
+        $this->authorize('update', $humanResource);
 
         $validated = $this->humanResourceService->validator($request, $id)->validate();
         try {
@@ -119,7 +136,7 @@ class HumanResourceController extends Controller
             ];
 
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
 
         return Response::json($response, ResponseAlias::HTTP_CREATED);
@@ -129,10 +146,12 @@ class HumanResourceController extends Controller
      * Remove the specified resource from storage.
      * @param int $id
      * @return Exception|JsonResponse|Throwable
+     * @throws AuthorizationException|Throwable
      */
     public function destroy(int $id): JsonResponse
     {
         $humanResource = HumanResource::findOrFail($id);
+        $this->authorize('delete', $humanResource);
 
         try {
             $this->humanResourceService->destroy($humanResource);
@@ -145,7 +164,7 @@ class HumanResourceController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
 
         return Response::json($response, ResponseAlias::HTTP_OK);
@@ -153,14 +172,15 @@ class HumanResourceController extends Controller
 
     /**
      * @param Request $request
-     * @return Exception|JsonResponse|Throwable
+     * @return JsonResponse
+     * @throws Throwable
      */
-    public function getTrashedData(Request $request)
+    public function getTrashedData(Request $request): JsonResponse
     {
         try {
             $response = $this->humanResourceService->getTrashedHumanResourceList($request, $this->startTime);
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response);
     }
@@ -168,9 +188,10 @@ class HumanResourceController extends Controller
 
     /**
      * @param int $id
-     * @return Exception|JsonResponse|Throwable
+     * @return JsonResponse
+     * @throws Throwable
      */
-    public function restore(int $id)
+    public function restore(int $id): JsonResponse
     {
         $humanResource = HumanResource::onlyTrashed()->findOrFail($id);
         try {
@@ -184,16 +205,17 @@ class HumanResourceController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response, ResponseAlias::HTTP_OK);
     }
 
     /**
      * @param int $id
-     * @return Exception|JsonResponse|Throwable
+     * @return JsonResponse
+     * @throws Throwable
      */
-    public function forceDelete(int $id)
+    public function forceDelete(int $id): JsonResponse
     {
         $humanResource = HumanResource::onlyTrashed()->findOrFail($id);
         try {
@@ -207,7 +229,7 @@ class HumanResourceController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response, ResponseAlias::HTTP_OK);
     }
