+++ app/Http/Controllers/HumanResourceTemplateController.php
@@ -6,6 +6,7 @@ use App\Models\HumanResourceTemplate;
 use App\Services\HumanResourceTemplateService;
 use Carbon\Carbon;
 use Exception;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Response;
@@ -44,14 +45,17 @@ class HumanResourceTemplateController extends Controller
      * @param Request $request
      * @return Exception|JsonResponse|Throwable
      * @throws ValidationException
+     * @throws AuthorizationException|Throwable
      */
     public function getList(Request $request): JsonResponse
     {
+        $this->authorize('viewAny', HumanResourceTemplate::class);
+
         $filter = $this->humanResourceTemplateService->filterValidator($request)->validate();
         try {
             $response = $this->humanResourceTemplateService->getHumanResourceTemplateList($filter, $this->startTime);
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response);
     }
@@ -59,13 +63,19 @@ class HumanResourceTemplateController extends Controller
     /**
      * @param int $id
      * @return Exception|JsonResponse|Throwable
+     * @throws AuthorizationException
+     * @throws Throwable
      */
     public function read(int $id): JsonResponse
     {
         try {
             $response = $this->humanResourceTemplateService->getOneHumanResourceTemplate($id, $this->startTime);
+            if (!$response) {
+                abort(ResponseAlias::HTTP_NOT_FOUND);
+            }
+            $this->authorize('view', $response['data']);
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response);
     }
@@ -74,9 +84,11 @@ class HumanResourceTemplateController extends Controller
      * @param Request $request
      * @return Exception|JsonResponse|Throwable
      * @throws ValidationException
+     * @throws AuthorizationException|Throwable
      */
     function store(Request $request): JsonResponse
     {
+        $this->authorize('create', HumanResourceTemplate::class);
 
         $validatedData = $this->humanResourceTemplateService->validator($request)->validate();
         try {
@@ -91,7 +103,7 @@ class HumanResourceTemplateController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response, ResponseAlias::HTTP_CREATED);
     }
@@ -101,10 +113,14 @@ class HumanResourceTemplateController extends Controller
      * @param int $id
      * @return Exception|JsonResponse|Throwable
      * @throws ValidationException
+     * @throws AuthorizationException|Throwable
      */
     public function update(Request $request, int $id): JsonResponse
     {
         $humanResourceTemplate = HumanResourceTemplate::findOrFail($id);
+
+        $this->authorize('update', $humanResourceTemplate);
+
         $validated = $this->humanResourceTemplateService->validator($request, $id)->validate();
         try {
             $data = $this->humanResourceTemplateService->update($humanResourceTemplate, $validated);
@@ -118,7 +134,7 @@ class HumanResourceTemplateController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response, ResponseAlias::HTTP_CREATED);
     }
@@ -127,10 +143,12 @@ class HumanResourceTemplateController extends Controller
      * Remove the specified resource from storage
      * @param int $id
      * @return Exception|JsonResponse|Throwable
+     * @throws AuthorizationException
      */
     public function destroy(int $id): JsonResponse
     {
         $humanResourceTemplate = HumanResourceTemplate::findOrFail($id);
+        $this->authorize('delete', $humanResourceTemplate);
         try {
             $this->humanResourceTemplateService->destroy($humanResourceTemplate);
             $response = [
@@ -142,7 +160,7 @@ class HumanResourceTemplateController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response, ResponseAlias::HTTP_OK);
     }
@@ -156,7 +174,7 @@ class HumanResourceTemplateController extends Controller
         try {
             $response = $this->humanResourceTemplateService->getTrashedHumanResourceTemplateList($request, $this->startTime);
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response);
     }
@@ -180,7 +198,7 @@ class HumanResourceTemplateController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response, ResponseAlias::HTTP_OK);
     }
@@ -203,7 +221,7 @@ class HumanResourceTemplateController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response, ResponseAlias::HTTP_OK);
     }
