+++ app/Http/Controllers/OccupationController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Exception;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 use App\Models\Occupation;
 use Carbon\Carbon;
@@ -44,14 +45,17 @@ class OccupationController extends Controller
      * @param Request $request
      * @return Exception|JsonResponse|Throwable
      * @throws ValidationException
+     * @throws AuthorizationException
      */
     public function getList(Request $request): JsonResponse
     {
+        $this->authorize('viewAny', Occupation::class);
+
         $filter = $this->occupationService->filterValidator($request)->validate();
         try {
             $response = $this->occupationService->getOccupationList($filter, $this->startTime);
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response);
     }
@@ -65,8 +69,12 @@ class OccupationController extends Controller
     {
         try {
             $response = $this->occupationService->getOneOccupation($id, $this->startTime);
+            if (!$response) {
+                abort(ResponseAlias::HTTP_NOT_FOUND);
+            }
+            $this->authorize('view', $response['data']);
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response);
     }
@@ -76,9 +84,12 @@ class OccupationController extends Controller
      * @param Request $request
      * @return Exception|JsonResponse|Throwable
      * @throws ValidationException
+     * @throws AuthorizationException
      */
     function store(Request $request): JsonResponse
     {
+        $this->authorize('create', Occupation::class);
+
         $validated = $this->occupationService->validator($request)->validate();
         try {
             $data = $this->occupationService->store($validated);
@@ -92,7 +103,7 @@ class OccupationController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response, ResponseAlias::HTTP_CREATED);
     }
@@ -103,12 +114,14 @@ class OccupationController extends Controller
      * @param int $id
      * @return Exception|JsonResponse|Throwable
      * @throws ValidationException
+     * @throws AuthorizationException
      */
     public function update(Request $request, int $id): JsonResponse
     {
         $occupation = Occupation::findOrFail($id);
-        $validated = $this->occupationService->validator($request, $id)->validate();
+        $this->authorize('update', $occupation);
 
+        $validated = $this->occupationService->validator($request, $id)->validate();
         try {
             $data = $this->occupationService->update($occupation, $validated);
             $response = [
@@ -121,7 +134,7 @@ class OccupationController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response, ResponseAlias::HTTP_CREATED);
     }
@@ -130,10 +143,12 @@ class OccupationController extends Controller
      * Remove the specified resource from storage
      * @param int $id
      * @return Exception|JsonResponse|Throwable
+     * @throws AuthorizationException
      */
     public function destroy(int $id): JsonResponse
     {
         $occupation = Occupation::findOrFail($id);
+        $this->authorize('delete', $occupation);
         try {
             $this->occupationService->destroy($occupation);
             $response = [
@@ -145,7 +160,7 @@ class OccupationController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response, ResponseAlias::HTTP_OK);
     }
@@ -159,7 +174,7 @@ class OccupationController extends Controller
         try {
             $response = $this->occupationService->getTrashedOccupationList($request, $this->startTime);
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response);
     }
@@ -171,7 +186,7 @@ class OccupationController extends Controller
      */
     public function restore(int $id)
     {
-        $occupation  = Occupation::onlyTrashed()->findOrFail($id);
+        $occupation = Occupation::onlyTrashed()->findOrFail($id);
         try {
             $this->occupationService->restore($occupation);
             $response = [
@@ -183,7 +198,7 @@ class OccupationController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response, ResponseAlias::HTTP_OK);
     }
@@ -194,7 +209,7 @@ class OccupationController extends Controller
      */
     public function forceDelete(int $id)
     {
-        $occupation= Occupation::onlyTrashed()->findOrFail($id);
+        $occupation = Occupation::onlyTrashed()->findOrFail($id);
         try {
             $this->occupationService->forceDelete($occupation);
             $response = [
@@ -206,7 +221,7 @@ class OccupationController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response, ResponseAlias::HTTP_OK);
     }
