+++ app/Http/Controllers/OrganizationTypeController.php
@@ -2,8 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Organization;
 use Carbon\Carbon;
 use Exception;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use App\Models\OrganizationType;
@@ -44,14 +46,17 @@ class OrganizationTypeController extends Controller
      * @param Request $request
      * @return Exception|JsonResponse|Throwable
      * @throws ValidationException
+     * @throws AuthorizationException|Throwable
      */
     public function getList(Request $request): JsonResponse
     {
+        $this->authorize('viewAny', OrganizationType::class);
+
         $filter = $this->organizationTypeService->filterValidator($request)->validate();
         try {
             $response = $this->organizationTypeService->getAllOrganizationType($filter, $this->startTime);
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response);
     }
@@ -65,8 +70,13 @@ class OrganizationTypeController extends Controller
     {
         try {
             $response = $this->organizationTypeService->getOneOrganizationType($id, $this->startTime);
+            if (!$response) {
+                abort(ResponseAlias::HTTP_NOT_FOUND);
+            }
+            $this->authorize('view', $response['data']);
+
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response);
     }
@@ -76,9 +86,12 @@ class OrganizationTypeController extends Controller
      * @param Request $request
      * @return Exception|JsonResponse|Throwable
      * @throws ValidationException
+     * @throws AuthorizationException
      */
     public function store(Request $request): JsonResponse
     {
+        $this->authorize('create', OrganizationType::class);
+
         $validated = $this->organizationTypeService->validator($request)->validate();
         try {
             $data = $this->organizationTypeService->store($validated);
@@ -92,7 +105,7 @@ class OrganizationTypeController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response, ResponseAlias::HTTP_CREATED);
     }
@@ -103,10 +116,13 @@ class OrganizationTypeController extends Controller
      * @param int $id
      * @return Exception|JsonResponse|Throwable
      * @throws ValidationException
+     * @throws AuthorizationException
      */
     public function update(Request $request, int $id): JsonResponse
     {
         $organizationType = OrganizationType::findOrFail($id);
+        $this->authorize('update', $organizationType);
+
         $validated = $this->organizationTypeService->validator($request, $id)->validate();
         try {
             $data = $this->organizationTypeService->update($organizationType, $validated);
@@ -120,7 +136,7 @@ class OrganizationTypeController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response, ResponseAlias::HTTP_CREATED);
     }
@@ -129,10 +145,13 @@ class OrganizationTypeController extends Controller
      * Remove the specified resource from storage.
      * @param int $id
      * @return Exception|JsonResponse|Throwable
+     * @throws AuthorizationException
      */
     public function destroy(int $id)
     {
         $organizationType = OrganizationType::findOrFail($id);
+        $this->authorize('delete', $organizationType);
+
         try {
             $this->organizationTypeService->destroy($organizationType);
             $response = [
@@ -144,7 +163,7 @@ class OrganizationTypeController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response, ResponseAlias::HTTP_OK);
     }
@@ -158,7 +177,7 @@ class OrganizationTypeController extends Controller
         try {
             $response = $this->organizationTypeService->getAllTrashedOrganizationUnit($request, $this->startTime);
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response);
     }
@@ -182,7 +201,7 @@ class OrganizationTypeController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response, ResponseAlias::HTTP_OK);
     }
@@ -205,7 +224,7 @@ class OrganizationTypeController extends Controller
                 ]
             ];
         } catch (Throwable $e) {
-            return $e;
+            throw $e;
         }
         return Response::json($response, ResponseAlias::HTTP_OK);
     }
